Store new plants in database. Fix LocationService subscribe/unsubscribe bug: {} -> ().

+++ app/src/main/java/com/geobotanica/geobotanica/data/dao/LocationDao.kt
@@ -7,8 +7,8 @@ import com.geobotanica.geobotanica.data.entity.Location
 @Dao
 interface LocationDao : BaseDao<Location> {
     @Query("SELECT * FROM locations WHERE id=:id")
-    fun get(id: Int): Location
+    fun get(id: Long): Location
 
     @Query("SELECT * FROM locations WHERE plantId = :plantId ORDER BY timestamp ASC LIMIT 1")
-    fun plantLocation(plantId: Int): Location
+    fun plantLocation(plantId: Long): Location
 }
